Animation Frame Tool 하나로 병합
- 이제 하나의 UI에서 처리

+++ 1. Project/Client/FrameTool.h
@@ -39,7 +39,7 @@ public:
     void SetAnimaton(DWORD_PTR _animKey);
 
 public:
-    void DeleteFrame(size_t& idx);
+    void DeleteFrame(int& idx);
 
 public:
     FrameTool();
